fix: rename slice to appStateSlice
